+++ phylopandas/__init__.py
@@ -1,3 +1,4 @@
-from . import phylopandas as _phylopandas
-from .dataframe import *
+from .frame import DataFrame
+from .read import *
+from .core import *
 from pandas import *
