+++ phylopandas/__version__.py
@@ -1,1 +1,1 @@
-__version__ ='0.1.3'
+__version__ = '0.1.4'
